machines: tqma[6,6ul,6ull,7]-mbaxx: work around meta-freescale overcleverness
meta-freescales tries to be smart but fails in default assignment of
IMAGE_BOOT_FILES from KERNEL_DEVICETREE. In imx-base.inc a string transformation
expects the list of dtb files to be single delimited which is not forced by bibake
language. Since we have no hard dependency from meta-freescale and have to assign
IMAGE_BOOT_FILES anyway, make the optional assignments in our <machine>.conf
a direct assignment. Hopefully we fixed the issue.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>