wic/tqmls1012al.wks: make device agnostic and prevent writing /etc/fstab
the --ondisk option requires the correct device name where a partition is
located. This will be written by a rootfs postprocess cmd into /etc/fstab.
When using systemd this leads to generating a mount unit, which will wait
for a long time and lead systemd start errors if this device is not present.

Remove the way it was done for other boards.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>