mc-utils: use meta-freescale version
Only used by tqmls1088a, which already requires meta-freescale.

To account for the changed deployed filenames, use a .wks.in file and
add MC_DPL and MC_DPC to WICVARS.

mc-utils_git.bb is copied from meta-freescale dunfell for now, as the
zeus version does not support the MC_FLAVOUR variable. The local copy of
this recipe can be removed after moving to Yocto dunfell.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>