tqmarzg2x: adapt arm-trusted-firmware from VLP_1.06 to TQMaRZG2x modules
And add support for QSPI-NOR-Flash, too.
The original implementation failed to setup the specified clock correctly.
Now clocks are ok and quad mode too.
Another bug fixed in ATF was that accessing the QSPI-NOR flash only worked
if the CPU booted from that device.

All patches, except the patches 010*.patch originate from Renesas and are
unmodified.

Signed-off-by: Andreas Dumberger <andreas.dumberger@tq-group.com>