conf/distro/tq-common.inc: set IMX_DEFAULT_BSP if using meta-freescale
meta-freescale uses imx-base.inc to set a lot of defaults for imx based
machines depending on MACHINEOVERRIDES. One of the base decisions is made
base on ${IMX_DEFAULT_BSP} which is use to construct the override
`use-${IMX_DEFAULT_BSP}-bsp:`

We set this to nxp here, if meta-freescale is used to make use of
features based on NXP drivers, firmware, binaries and forks.

For our tiny spaetzle distro we force to linux mainline based kernel
in this case.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>