imx-sc-firmware: rename recipe to have own flavour
Since we modified the logic to fit it to our needs, rename the recipe
to imx-sc-firmware-tq, but provide imx-sc-firmware with the recipe.
This ensures that the correct recipe is used without depend on ugly
things in the bblayers order and / or layer priority.

While doing so, we need also to set the PREFERRED_PROVIDER in machine
files to keep it working.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>