imx-boot: port bootstream generator from zeus-tqma8
- sync with recipes in meta-freescale
- keep multi bootstream build feature
- use new recipe name

Since we modified the logic to fit it to our needs, rename the recipe
to imx-boot-tq, but provide imx-boot with the recipe. This ensures
that the correct recipe is used without doing things in the bblayers
order and / or layer priority.

While doing so, we need also to set the PREFERRED_PROVIDER in machine
files to keep it working.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>