tqma65xx-mba65xx: add machine configurations
There are two configuration variants for different RAM configurations of
the TQMa65xx ("2g" and "4g"). In addition, machine configurations for the
R5 core of the AM65xx are added as multiconfig targets to build the R5
bootloader, which is required for the first stage of the boot.

Separate configurations are needed for the two revisions of the AM65xx
(k3r5 and k3r5-sr2) for compatiblity with the machine configuration in
meta-ti that is also used for the AM65xx EVM. The tqma65xx always uses SR2.

Our configuration deviates from the meta-ti defaults by using a separate
TMPDIR for each multiconfig, to allow the build to succeed with rm_work.
The two k3r5 machines still share most of their build results, in
particular the R5 toolchain, through the sstate cache.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>