hdparm: add include for sys/sysmacros.h
The macros minor, major and makedev are only defined in sys/sysmacros.h
in glibc since v2.28. See commit e16deca62e16 ([BZ #19239] Don't include
sys/sysmacros.h from sys/types.h) in
https://sourceware.org/git/?p=glibc.git.

Because the rzg2 layer uses cip, this glibc v2.28 is used instead of
v2.26 from poky. This is not compatible with hdparm-9.52. To fix this,
sys/sysmacros.h is included in sysfs.c and geom.c.

This change is also part of hdparm-9.55.

Signed-off-by: Paul Gerber <Paul.Gerber@tq-group.com>