tq-image-small: install some more essential packages
kernel and device trees:

Since the image is intended to be used for UBI on QSPI we
install these packages in the rootfs. This will save space.
Besides this, kernel and modules needs to be updated together
anyway.

packagegroup-netutils:

All our starterkits have network. So let's install some
essential tools.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>