conf/distro: add intermediate includes for spaetzle and dumpling
Depending on the distro we want to switch between different kernel
provider and maybe depending on that also graphic / multimedia stacks.

We separated the provider logic in the past but used the distro file as
include to build vendor specific flavours. When doing only partial
overrides of some settings, this can produce hard to find build issues.

Separate the basic settings for a distro family and use them as include
to be used by the different flavours.

Signed-off-by: Markus Niebel <Markus.Niebel@tq-group.com>