machine/tqmxx: Set non-machine specific PREFERRED_PROVIDER_virtual/kernel
Instead of setting PREFERRED_PROVIDER_virtual/kernel for a machine
specific override, set the value unrestricted.
This allows overriding the provider in a local.conf without using this
specific override.
Update the tqmls1028a documentation as well which shows the benefit.
This is also a preparation for changing the provider on distro level.

Signed-off-by: Alexander Stein <alexander.stein@tq-group.com>