recipes-kernel: add port of firmware-imx recipe from meta-freescale
This port is intended to be used as a replacement for firmware-imx when
building for TQ-Systems i.MX SOM without meta-freescale. Main reason is
to enable to build mainline based systems without lots hard to oversee
stuff from vendor driven layer.

To prevent conflicts when using with meta-freescale RPROVIDES, RCONFLICTS
and RREPLACES are set for the created subpackages.

This recipe uses the newly added nxp-bin-unpack class.

At the moment i.MX6 VPU firmware is supported.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>