tq-image-debug: extend etc/issue in ROOTFS_POSTPROCESS_COMMAND
Adjust /etc/issue[.net] in a separate postprocess function
if IMAGE_FEATURES contains debug settings. Has to be done this
way, to have access to IMAGE_FEATURES variable - it is not
guaranteed that the final value of IMAGE_FEATURES is seen
by recipes for regular packages.

While at it, fix typo: poke -> poky

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>