recipes-support/rng-tools: Do not start if HW RNG is available
Modify rngd.service that it is only started if 'ExecCondition' returns 0,
thus no HW RNG is available. If check_hwrng returns 1 rngd.service is
skipped, thus marked as 'inactive' and _not_ failed.
Rationale is that rngd is only useful if no HW RNG is available and
libjitterentropy provides entropy during bootup. If HW RNG is available
[hwrng] kthread does this job already.

Signed-off-by: Alexander Stein <alexander.stein@tq-group.com>