conf/layer.conf: Ignore dynamic-layers/baser-imx8 when meta-imx is used
The dynamic-layer recipes are for providing some base .bb when
meta-imx resp. meta-bsp is not added. But as meta-tq has a higher priority
over meta-imx the original files from meta-imx are still overriden by
meta-tq.
In order to provide the base .bb files _only_ when meta-imx is not added,
ignore the dynamic overrides when meta-imx is added to BBFILE_COLLECTIONS

Signed-off-by: Alexander Stein <alexander.stein@tq-group.com>