ti: wic: use --fixed-size, increase boot partition size
Starting with dosfstools 4.2, mkfs.vfat will align the created
filesystem to certain sector numbers. This can cause the filesystem to
be smaller than the containing partition, which causes boot failures on
the TQMa57xx.

Use --fixed-size to ensure that the created partition always has an
aligned sector number, and increase the size to make space for newer
kernels. Also replace --align with the new --offset option for the
position of the partition.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>