dynamic-layers/freescale-layer: rcw: remove local OVERRIDE
The recipe adds a local OVERRIDE to replace the source repo
with the TQ-Systems clone of the rcw repo. Since replacing
the source URI for a recipe is a normal reason for a bbappend
there is no need for this type of OVERRIDE usage. The new URI
references a fork of the original repo, so all other boards
from meta-freescale can still be built.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>