layer.conf: add dynamic-layer handling for meta-oe
This is needed if we have lmsensors + sensors-config in our images.
The bbapends were ported but adding to dynamic layers was left out
when integrating tqma8 in hardknott branch.

While at it, fix BBFILES_DYNAMIC to use 4 space indentation, since
this is what we do in multiline assignments.

Backport from kirkstone-next.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>