lmsensors-config: improve conversion factors for MBa8MPxL
Use correct value of 7.8 for input resistor network. Note that this also
needs a dtb with the correct settings for ADC reference voltage.

While at it use wildcard for the chip name and add some comments to
make clear that the files use is solely for MBa8MPxL.

Backport from kirkstone-next

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>