u-boot-ti-tq-2021.01: update for TQMa64xxL/MBaX4XxL REV 0200
In addition, the new version includes the latest ti-u-boot upstream,
with a number of new bugfixes and errata workarounds.

List of changes excluding upstream commits:

    11eb6e967d1 board: tqc: tqma64xxl: enable SD card UHS mode support
    30f18fc4608 arm: dts: k3-am642-tqma64xxl-mbax4xxl: update DTS from kernel
    22c47a0e65c arm: dts: k3-am64: move VTM node to main domain
    9494e8d3698 arm: dts: k3-am64: Add VTM node
    4916f1e6a7a board: tqc: tqma64xxl: add netdev to default environment
    3bebbbf5e82 Merge remote-tracking branch 'ti/ti-u-boot-2021.01' into TQMaxx-ti-u-boot-2021.01-mr
    8274bdffbaf configs: tqma64xxl_mbax4xxl: update for latest ti-u-boot changes
    ae63bfeb6e8 Merge remote-tracking branch 'ti/ti-u-boot-2021.01' into TQMaxx-ti-u-boot-2021.01-mr
    f038782e40d Revert "arch: arm: mach-k3: am642_init: add reset workaround for CPSW hang"

Due to the merge of the GP and HS configurations in ti-u-boot, the
SPL_BINARY and UBOOT_BINARY variables need to be adjusted to always look
for files with _HS suffix, even in unsecured builds.

The new U-Boot version also requires a recent ti-sci-fw, which is
provided by meta-ti kirkstone 2b11afa8aa19814264d1d7b97184ad5cb88ddfa0
or later.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>