ti-sgx-ddk-km: fix build with kernel 5.10
Make the override of the driver branch conditional on the kernel version
and select an alternative set of patches when the version is set to 5.10.

As there is no nice way to determine the selected kernel version early
enough to affect SRC_URI, introduce a new variable TI_SGX_DDK_KM_KERNVER,
which machine or distro configs can override. We still default to
kernel 5.4, as all machines in meta-tq that use this driver
(TQMa335x/57xx/65xx) are based on this kernel version for now.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>