meta-tq: tqmxx.inc: implement IMAGE_BOOT_FILES helper for dtb
This is just a copy from meta-freescale that allows to convert a list
of dtb[o] into a list that fits into IMAGE_BOOT_FILES without vendor
directory prefix as currently used for aarch64.

This was intentionally copied to enable reuse for platforms where
such a functionality is not provided yet.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>