meta-tq: tqma8mxml-mba8mx: use new create_image_boot_dtb_files
This implements the same logic for `KERNEL_DEVICETREE` as for armv7a
and removes the need for an additional set of variables and
reimplementing the already existing logic.

Move tqma8-common.inc to tqma8mxml.inc since we use the functions
after `KERNEL_DEVICETREE` and `KERNEL_DEVICETREE_<provider>` are
assigned.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>