meta-tq: tqma8x[d,q]ps: use new create_image_boot_dtb_files
This implements the same logic for `KERNEL_DEVICETREE` as for armv7a
and removes the need for an additional set of variables and
reimplementing the already existing logic.

tqma8-common.inc is already included via tqma8xxs.inc including
tqma8xx-common.inc an can be removed  from machine conf files.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>