tqma64xxl: update machines for latest meta-ti
Latest meta-ti added R5 multiconfigs for 3 different variants of the AM64x
SoC, selecting one of 3 different SYSFW versions:

- GP (General Purpose)
- SR2.0 HS-FS (Field Securable)
- SR2.0 HS-SE (Security Enforced)

Mirror the AM64x-EVM machine configs for TQMa64xxL/MBaX4XxL to allow
building again.

Also add a flag to change the variant that is installed to the WIC image
as tiboot3.bin. For now we default to the GP variant to allow booting on
existing TQMa64xxL hardware.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>