ti-sgx-ddk-um: replace initscript with udev rule
The ti-sgx-ddk driver requires an additional userspace initialization
step after the kernel module has probed the device. Without this
initialization, no EGL context can be created and Weston will fail to
start.

The driver package contains an init script, but this does not work on
systemd-based systems. Introduce an enabled-by-default PACKAGECONFIG that
installs a udev rule instead to run the init command automatically when
the driver is loaded.

udev reports several events when the pvrsrvkm module is loaded:

- add event for the kernel module
- add events for two DRM devices
- bind event for the GPU platform device

The DRM devices aren't nice to match on, and the kernel module add is
too early to run `pvrsrvctl --start`, so we trigger on the platform
device bind.

With this change, Weston works out-of-the-box on the MBa65xx.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>