meta-tq: tqma8-common.inc: allow BOARD_DEVICETREE_FILES to be empty
The variable may be empty / not set in case KERNEL_DEVICETREE is set
including the vendor prefix.

This is a transitional fix and the complete usage of the variable and
the anonymous python function will be removed after all boards are
converted to the new logic that uses create_image_boot_dtb_files and
kernel_provider_dtbs.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>