meta-tq: fix WKS_FILE_DEPENDS for ti machines
When creating a wic image with ext4 rootfs, the wic logic will
create the ext4 on its own. Add e2fsprogs-native to
WKS_FILE_DEPENDS the same way we added dosfstools-native.
This guarantees that the tools are there even if
no separate ext4 image will be built.

This fixes errors on massive parallel builts wic creation
sometimes fails with error for missing e2fsprogs-native packages
with a hint to add this to the wic-tool dependencies, where this already is.

While at it, reorder the WKS_FILE_DEPENDS alphabetically and move
targets that could also be used for non wic targets to EXTRA_IMAGEDEPENDS
like it was already done in commit
15c121a8 ("meta-tq: conf/machine: improve EXTRA_IMAGEDEPENDS / WKS_FILE_DEPENDS")

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>