meta-tq: qoriq-atf: add support for tqmls10xxa machines
In contrast to chassis 3 qoriq modules (TQMLS1028A or TQMLX2160a) chassis 2
qoriq modules (TQMLS1043a and TQMLS1046a) have the bootsource hardcoded in
the rcw file. So when generating the different variants we do not have to
consider configurations where the mode of the rcw and the bootmode do not match.

When the rcw filename contains a bootsource build BL2 PBL for this bootsource
only.

Signed-off-by: Gregor Herburger <gregor.herburger@tq-group.com>