meta-tq: tqmls104?a-mbls10xx: update to new atf bootstream
Previously the bootstream for the tqmls10xx modules was from bootrom ->
u-boot -> linux. For secure boot ppa had to be loaded in u-boot. The rcw
binary was generated in U-Boot and concatenated to the U-Boot binary.

Because ppa software is deprecated move to new bootstream (bootrom ->
atf (arm trusted firmware) -> u-boot -> linux.

Also generate the rcw binary with the rcw recipe which is than used by
the qoriq-atf recipe to generate the pbl binary. U-Boot is updated to
new version to support atf bootstream.

The machine configurations for tqmls1043a_2g-mbls10xxa,
tqmls1046a_8g-mbls10xxa and tqmls1046a-noecc-mbls10xxa are removed
because variants for this hardware configurations are now build with the
qoriq-atf recipe.

Signed-off-by: Gregor Herburger <gregor.herburger@tq-group.com>