README.md, doc: list TQMa6442L and MBaX4XxL REV.010x as supported
As we support the old revisions for now, also add a few hardware issues to
the list of known issues.

No issues with the SD card detection have been observed in practice even
when redetecting a card after switching to UHS mode, but it is still
listed as a potential issue for completeness.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>