imx-boot-tq: mx8m: generate habinfo file
Parse the output of imx8_mkimage and print_fit_hab to collect CSF
offsets and the HAB block list.

This will create a file with content of the following format for each
imx-boot.bin variant:

    SPL_CSF_OFF="0x29000"
    SPL_HAB_BLOCK="0x911fc0 0x0 0x29000"
    SLD_CSF_OFF="0x59020"
    SLD_HAB_BLOCK="0x401fcdc0 0x58000 0x1020"
    FIT_HAB_BLOCK="0x40200000 0x5B000 0xB3268
    0x402B3268 0x10E268 0x74E8
    0x960000 0x115750 0xB1B0
    "

With this information, signatures for the bootstream can be generated
and patched in, allowing to separate the build host from the signing
keys.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>