meta-tq: define tqma8mxml-multi-mba8mx
This machine uses the improved wic logic to build images
for different RAM configs.

This way we can use a single MACHINE instead of numbers of
different MACHINEs to build all needed bootloader flavours and
generating a set of bootable disk images. This will help reducing
build time and space.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>