meta-tq: add class supporting generation of additional wic images for a machine
This class generates additional wic images for machines defining list variable
'EXTWIC_BOOTLOADER_IMAGES'. This is useful for boards differing for instance
only in the assembled RAM. Technically only the boot image in the unpartitioned
area of the wic image will be replaced.

To use the class, RAW_BOOT_END_OFFSET_KB and RAW_BOOT_START_OFFSET_KB have to be
defined for the machine using the class and should be used for placement of the
alternate boot images in WIC files.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>