meta-tq: tq-extlinux: new recipe
tq-extlinux creates an extlinux.conf based on the uboot-extlinux-config
bbclass and installs it to /boot/extlinux/.

The generated configuration may look like the following:

    # Generic Distro Configuration file generated by OpenEmbedded
    LABEL Dumpling (TQ-Systems Dumpling Distribution) for TQ-Systems NXP BSP
        KERNEL ../Image
        FDTDIR ../
        APPEND ${bootargs_root} rootwait rw console=${console},${baudrate}

The package is basically equivalent to the <u-boot>-extlinux packages
provided by every U-Boot recipe. As the generated file does not actually
depend on U-Boot, it is more convenient to have a separate recipe with a
fixed name to refer to however.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>