meta-tq: tq-bootscripts: new recipe
Add recipe to build boot.scr images with U-Boot command scripts.

For now, we provide a single script "boot-blockdev", which sets up the
root= cmdline for eMMC/SD-card (with a fallback for other block devices)
and then boots using an extlinux.conf in the rootfs.

The FIT format is chosen for boot.scr, so the same build code can be
used for Secure Boot by adding FIT image signatures.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>