meta-tq: tqmlx2160a.wks.in: use offset instead of align
For absolute placement use offset. While at it, add the non partitioned
region for U-Boot env and fix related documentation according to actual
U-Boot settings for CONFIG_ENV_OFFSET.
Later on this can be extended to write environment image(s) to the WIC-image.
Remove --fixed-size for /boot and adjust layout.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>
Signed-off-by: Gregor Herburger <gregor.herburger@ew.tq-group.com>