meta-tq: dynamic-layers: freescale: mc-utils: add PHY
With update to Linux version v6.1 the XFI interface on mac 13 and 14
now supports SFP modules. Although this is not a real PHY the link_type
must be MAC_LINK_TYPE_PHY to support the phylink interface.

While at it set all other used macs to MAC_LINK_TYPE_PHY as there is
always a real phy connected.

Signed-off-by: Gregor Herburger <gregor.herburger@tq-group.com>