Revert "mc-utils: dpc-warn.dts: fix link_type"
With commit 250f4fb175fcd link_type was set to MAC_LINK_TYPE_FIXED
because of problems with RGMII on MAC 17 and 18. Actually
MAC_LINK_TYPE_PHY is correct because on these macs a phy is connected in
all configurations (rgmii or sgmii).

In older versions of the dpaa2 firmware the precedence of rgmii over sgmii
on port 17 and 18 was wrong. Presumably this caused the problem.

With newer versions (tested with 10.35.0) of dpaa2 firmware the problem could
not be reproduced so change back to MAC_LINK_TYPE_PHY.

This reverts commit 250f4fb175fcd4f7834be35a68fe094f5a79a0fd.

Signed-off-by: Gregor Herburger <gregor.herburger@ew.tq-group.com>