meta-tq: doc: README.TQMT10xx.md: various fixes
In the tables showing the DIP-switch settings the function description
does not match the correct DIP-Switch.
S2-4 selects the reference clock for the Serdes Module on the SoM.
S9-1 selects the reference clock for the QSGMII Ethernet Phy.

Correct this and rename ambigous `Refclk 125 Mhz` to `Refclk Serdes`.

Also wrong port names are fixed.

Signed-off-by: Gregor Herburger <gregor.herburger@tq-group.com>