meta-tq: TQMa8/TQMa9: use RAW_BOOT_END_OFFSET_KB / RAW_BOOT_START_OFFSET_KB for wic
To support sharing the WKS_FILE but having different offset for boot images
and to prepare using a class allow to build multiple wic-images for a single
machine using different boot images we need to define variables for
the placements of said images.

Since IMX_BOOT_SEEK is defined for imx8 / imx9 by meta-freescale for
the placement, keep this variable defined for compatibility.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>