meta-tq: ti-sgx-ddk-km: fix build with latest meta-ti
The ti-sgx-ddk-km recipe has changed significantly with the latest
updates in the source repository, so simply switching back to an older
version in a bbappend is not possible anymore.

Instead, the following approach is implemented:

- The old recipe is imported into meta-tq as a separate recipe
- Either the old or the upstream recipe is skipped based on the value or
  TI_SGX_DDK_KM_KERNVER

As the recipe doesn't directly depend on the one in meta-ti-bsp anymore,
it can work with both old and new versions of the meta-ti repository.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>