linux-tq: make localversion deterministic
The localversion string built into the kernel should be deterministic.
With OpenEmbedded it is common practice to apply patches to the sources.
The commit id of HEAD will differ each time the patching step is rerun.
With the previous implementation of the do_set_localversion_tq task this
produced non-deterministic commit ids to be used. This is undesireable,
because the commit id read from the version string during runtime is
not present in the sources. It also caused problems when building out-
of-tree kernel-modules if the kernel itself was taken from sstate cache
but the kernel-module needed to be rebuild and retriggered the kernels
do_patch task (see this issue in meta-freescale, that was due to a
similar localversion modification
https://github.com/Freescale/meta-freescale/issues/961).

Rework to produce a deterministic string that will use the commit id
given as SRCREV and count the number of patches applied on top.

With e.g. commit id 11aabbcc and 5 patches this will result in a version
string
  +g11aabbcc+p5

Related: https://github.com/Freescale/meta-freescale/pull/1694

Signed-off-by: Daniel Wagenknecht <dwagenknecht@emlix.com>