meta-tq: machine: set IMAGE_FSTYPES for all machines
Various include snippets from vendor layers set IMAGE_FSTYPES to
different values that are not useful for TQ modules. Instead of relying
on setting IMAGE_FSTYPES in local.conf or a similar configuration,
ensure that the defaults for each machine make sense.

Instead of using append/remove, which made it difficult to modify the
settings in derived machines and distro configs in the past, the
variable is set using a regular assignment.

In addition to the .wic and .tar.gz, this adds also the .wic.bmap (block
map), so bmaptool can be used to program images more efficiently.

The default setting in the distro becomes redundant an can be removed.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>