meta-tq: dynamic-layers/freescale-layer: imx-mkimage_git: add bbappend
At the time we forked imx-boot recipe we missed to update SRC_URI and
SRCREV for imx-mkimage tooling. Use bbappend to the recipe of the tools
to support building the tools with the same set of patches and
configurations as it is used for the recipe used to build the boot
images itself.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>