linux-firmware: add fix for upgrade to 20231030
Image size increased for linux-firmware-ath10 due to new
dependency to linux-firmware-qcom-sdm845-modem. Since this
is noted as a backward compatibility which we do not need,
drop it.

This fix may need adjustments when using with newer linux-firmware
versions or when integrated in upstream yocto recipe.

Fixes poky commit a818202d85a2 ("linux-firmware: upgrade 20230804 -> 20231030")

While at it, remove a superfluous space in assignment to SRC_URI.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>