meta-dumpling: distro: tq-provider-ls: fix IMAGE_FSTYPES overrides
`IMAGE_FSTYPES:override` will completely replace `IMAGES_FSTYPES`, even
when `+=` is used. Assign the additions to another variable first to
avoid the override of `IMAGE_FSTYPES`.

It might be better to move this configuration to the machine
configuration to avoid using overrides at all, but for now we just fix
the bug and leave further restructuring for later.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>