tqma6ullx[l]: prepare multi machine config
This machine uses the improved wic logic building images
for different RAM configs and form factors.

This way we can use a single MACHINE instead of numbers of
different MACHINEs to build all needed bootloader flavours and
generating a set of bootable disk images. This helps to reduce
build time and space.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>