meta-dumpling: distro: spaetzle: fold poky-tiny into distro definition
Follow suggestions from yocto wiki for poky-tiny and use poky-tiny as a base
for defining an own slim down distribution. Since there are hard assumptions
for kernel, image types and the like in poky-tiny this approach will save
nasty workarounds for this. Besides said workarounds did not work well when
overriding from local.conf.

Add some comments in the distro definition that are shall help when
deriving customized distros.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>