meta-tq: imx-cst-keys: change keys and certificates for AHAB
The previously distributed fuse table using 256 bit is not compatible with
TQMa8x and TQMa8Xx, which need 512 bit to be fused. A AHAB fuse table
can only be generated with all 4 SRKs.

Provide a new set of example keys and certificates that can be used for
all SoCs with AHAB support. Additionally, a SoC specific fuse table is
provided.

Signed-off-by: Paul Gerber <Paul.Gerber@tq-group.com>