meta-tq: linux-tq.inc: improve set_localversion_tq task
Beginning with linux 6.3 information in '.scmversion' is not longer added
to the kernel version string. When building from git the information is
generated from SCM if CONFIG_LOCALVERSION_AUTO is enabled.

When using kernel recipes with local patches this has potential
issues:

- when git is used to apply patches the patches are considered as
  part of kernel tree, Everytime the pathes gets reapplied the commit
  hash changes an so do the kernel version string
- Using set_localversion_tq and CONFIG_LOCALVERSION_AUTO the commit
  information is added twice

Additionally yocto uses now 'KERNEL_LOCALVERSION'. Adding this to
the generated localversion-tq file will also apply this twice to the
version string. Do not use it in set_localversion_tq.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>