meta-dumpling: classes: add tq-buildinfo
The new class tailors our needs to integrate build information into the resulting
rootfs image and to put the same information in a readable file beneath the binary artifacts
using the 'IMAGE_POSTPROCESS_COMMAND' logic.

The existing 'image-buildinfo' is configured using a default set of variables added to
'IMAGE_BUILDINFO_VARS'. Based on our build space layout the revision of the build space
repository is queried and added.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>