meta-tq: machines: centralize MACHINE_EXTRA_RRECOMMENDS
All machines in meta-tq layer support device tree and allow building with
kernel modules.

Because we are about to move kernel and devicetree to /boot in rootfs
and all machines install these packages move this to shared include
meta-tq/conf/machine/include/tq-machine.inc.

To make overrides easier this new include shall be used at baseboard level
and not on module level.

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>