meta-tq: dynamic-layers: move and rewrite imx-oei
The solution proposed in meta-freescale has some drawbacks:

- When referencing the include to write an own recipe a full path
  must be given since the include is under dynamic layers in
  meta-freescale
- doing a bbappend here under dynamic layer for arm-toolchain may
  cause build errors, when arm-toolchain is present but meta-freescale
  where the recipe comes from is not

The proposed fix ist to add a check for arm-toolchain to the generic
include and raise a bb.parse.SkipPackage exception if not. Since
the recipe is only processed if in dependency chain or directly
addressed no build error is expected under normal conditions.

f3a45262f563 ("meta-tq: imx-oei: add TQ-specific .bbappend")

Signed-off-by: Markus Niebel <Markus.Niebel@ew.tq-group.com>