meta-tq: tq-image.inc: compress rootfs images using zstd
XZ compression is very slow, which is particularly annoying during
development. While Zstd's compression rate is a bit worse, it is
extremely fast, making it a better default (speeding up image compression
from ~90s to <1s for an ext4 image with ~350MiB of allocated blocks).

A TQ_IMAGE_COMPRESSION variable is introduced to make it easier to
switch the compression algorithm.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>