meta-tq: tqmls1088a, tqmlx2160a: install DPAA2 config (DPC, DPL) to rootfs
Prepare for loading the DPL from the rootfs instead of the boot
partition in U-Boot. This will allow for more robust updates/rollbacks
in A/B partition configurations.

The Management Complex firmware and DPC are needed for Ethernet support
in U-Boot, so they will continue to be loaded from the boot partition
for the time being.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>