meta-tq: conf: tq-image: add default logic for KERNEL_IMAGETYPES
Set KERNEL_IMAGETYPES to include KERNEL_IMAGETYPE and fitImage.
KERNEL_IMAGETYPES_INSTALL ensures that only KERNEL_IMAGETYPE is
installed for tiny builds. A custom mainboard can always override
KERNEL_IMAGETYPES as needed.

Not all built fitImages are necessarily usable, as UBOOT_LOADADDR and
UBOOT_ENTRYPOINT need to be set to a sensible value. This needs to be
fixed for each SoM individually.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>