meta-tq: machine: tqmlx2160a-mblx2160a: provide tfa-secure-boot U-Boot config
qoriq-atf expects to find a u-boot.bin-tfa-secure-boot when the `secure`
distro feature is enabled. In the future, a proper secured configuration
and default environment could be provided; for now, we just use the same
U-Boot config as for the non-secure build as an example for setting up a
secured boot flow until U-Boot, but not further.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>